requirement.txt
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0